Fixed arrayPos doing undefined behaviour
Fixed a bug causing arrayPos to do undefined behaviour. For some reason, the calibration() function causes both arrayPos and calibratedDistance to have the same value in the last loop of for loop. Don't know what causes this, but assigning it the value it should be after the loop seems to fix this.

Also moved arrayPos incrementation to it's own function called arrayPosition().